[Refactor] moving class associated to MutRepair to the corresponding package
